aidetic_Assifnment/consumer.py
#Data Ingestion from Kafka:

from kafka import KafkaConsumer

# Set up Kafka consumer
consumer = KafkaConsumer('clickstream_topic', bootstrap_servers='localhost:9092')

# Iterate over Kafka messages
for message in consumer:
    # Extract fields from the message
    user_id = message.value['user_id']
    timestamp = message.value['timestamp']
    url = message.value['url']
    ip_address = message.value['ip_address']
    user_agent = message.value['user_agent']

    # Store the extracted data in the data store
    # (code for storing the data in the chosen data store)
