aidetic_Assifnment/requirements.txt
kafka-python
pyspark
elasticsearch